requirements.txt
Flask
Flask-MySQLdb
python-dotenv
requests
gunicorn